Remove bad integer casts

+++ cas-server-core-tickets/src/main/java/org/jasig/cas/ticket/registry/DefaultTicketRegistry.java
@@ -89,12 +89,12 @@ public final class DefaultTicketRegistry extends AbstractTicketRegistry {
 
     @Override
     public long sessionCount() {
-        return (int) this.cache.values().stream().filter(t -> t instanceof TicketGrantingTicket).count();
+        return this.cache.values().stream().filter(t -> t instanceof TicketGrantingTicket).count();
     }
 
     @Override
     public long serviceTicketCount() {
-        return (int) this.cache.values().stream().filter(t -> t instanceof ServiceTicket).count();
+        return this.cache.values().stream().filter(t -> t instanceof ServiceTicket).count();
     }
 
     @Override
